+++ Signal/Pulsar/dsp/CovarianceMatrixCUDAEngine.h
@@ -60,7 +60,7 @@ __global__ void applyScaleKernel(float* amps, unsigned int ampsLength, double sc
 __global__ void genericAddKernel(unsigned int n, float* original, const float* add);
 __global__ void genericAddKernel(unsigned int n, unsigned int* original, const unsigned int* add);
 __global__ void genericDivideKernel(unsigned int n, float* d_numerators, unsigned int denominator);
-__global__ void checkForZeroesKernel(float* d_hits, unsigned int hitsLength, bool* d_zeroes);
+__global__ void checkForZeroesKernel(unsigned int* d_hits, unsigned int hitsLength, bool* d_zeroes);
